Revert "Add the with-package macro"
This reverts commit f94fef651ef9b21105c33b0eb2f14c80b88e6561.

I wanted to use with-package (a replacement for eval-after-load that can
take multiple features to depend on) to delay executing certain code
until all the packages that it depends on are loaded. It turned out that
with-package waits until any one of the specified features is provided
and requires the rest immediately.